% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediastoredata_operations.R
\name{delete_object}
\alias{delete_object}
\title{Deletes an object at the specified path}
\usage{
delete_object(Path)
}
\arguments{
\item{Path}{[required] The path (including the file name) where the object is stored in the container. Format: <folder name>/<folder name>/<file name>}
}
\description{
Deletes an object at the specified path.
}
\section{Accepted Parameters}{
\preformatted{delete_object(
  Path = "string"
)
}
}

