% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediastoredata_operations.R
\name{describe_object}
\alias{describe_object}
\title{Gets the headers for an object at the specified path}
\usage{
describe_object(Path)
}
\arguments{
\item{Path}{The path (including the file name) where the object is stored in the container. Format: <folder name>/<folder name>/<file name>}
}
\description{
Gets the headers for an object at the specified path.
}
\examples{

}
