% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediastoredata_operations.R
\name{get_object}
\alias{get_object}
\title{Downloads the object at the specified path}
\usage{
get_object(
  Path = "string",
  Range = "string"
)
}
\arguments{
\item{Path}{[required] The path (including the file name) where the object is stored in the container. Format: <folder name>/<folder name>/<file name>

For example, to upload the file \code{mlaw.avi} to the folder path \code{premium\canada} in the container \code{movies}, enter the path \code{premium/canada/mlaw.avi}.

Do not include the container name in this path.

If the path includes any folders that don't exist yet, the service creates them. For example, suppose you have an existing \code{premium/usa} subfolder. If you specify \code{premium/canada}, the service creates a \code{canada} subfolder in the \code{premium} folder. You then have two subfolders, \code{usa} and \code{canada}, in the \code{premium} folder.

There is no correlation between the path to the source and the path (folders) in the container in AWS Elemental MediaStore.

For more information about folders and how they exist in a container, see the \href{http://docs.aws.amazon.com/mediastore/latest/ug/}{AWS Elemental MediaStore User Guide}.

The file name is the name that is assigned to the file that you upload. The file can have the same name inside and outside of AWS Elemental MediaStore, or it can have the same name. The file name can include or omit an extension.}

\item{Range}{The range bytes of an object to retrieve. For more information about the \code{Range} header, go to \url{http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.35}.}
}
\description{
Downloads the object at the specified path.
}
\examples{

}
