% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediastoredata_operations.R
\name{list_items}
\alias{list_items}
\title{Provides a list of metadata entries about folders and objects in the specified folder}
\usage{
list_items(Path = NULL, MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{Path}{The path in the container from which to retrieve items. Format: <folder name>/<folder name>/<file name>}

\item{MaxResults}{The maximum results to return. The service might return fewer results.}

\item{NextToken}{The \code{NextToken} received in the \code{ListItemsResponse} for the same container and path. Tokens expire after 15 minutes.}
}
\description{
Provides a list of metadata entries about folders and objects in the specified folder.
}
\examples{

}
